knob_ctl: first commit
Add the knob_ctl classes for low-level manipulation of the etch-a-sketch.
Right now this just contains a dummy implementation that manipulates an
image in memory and writes it out to a file. This will allow testing of
the number writing code before the hardware works and easy replacement.

Right now this compiles as an executable just for testing. This will change
later, as it should just be an object file.